Enhance user model with location fields and update admin/user panels
- Add district, sub-district, address, and number fields to User model
- Change balance field type from bigInteger to decimal in migration
- Update admin and user panel primary colors to green
- Add sample address and balance values in UserSeeder
- Define district and sub-district relationships in User model
- Create new Filament directory structure for resources

+++ database/migrations/0001_01_01_000003_create_users_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger('district_id'); 
             $table->unsignedBigInteger('sub_district_id'); 
             $table->string('address')->nullable();
-            $table->bigInteger('balance')->default(0);
+            $table->decimal('balance')->default(0);
             $table->timestamps();
 
             $table->foreign('district_id')->references('id')->on('districts')->onDelete('cascade');
